services/bot/requirements.txt
twython
firebase-admin
google-cloud-firestore
scheduler
